requirements.txt
google-genai>=0.5.0
pdfplumber>=0.11.0
reportlab>=4.1.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
